outsource _apply_setops to module
